requirement.txt
GitPython==0.3.6
PySide==1.2.2
gitdb==0.6.4
smmap==0.9.0
wsgiref==0.1.2
